aider: Add 'first_move' block to the toolbox in index.html.
# Aider chat conversation:

USER: also update index.html

ASSISTANT: To include the new 'first_move' block in the toolbox of the Blockly interface, we need to modify the `client/index.html` file.

Here is the *SEARCH/REPLACE block* for the changes:

```html
client/index.html
<<<<<<< SEARCH
    <xml id="toolbox" style="display: none">
        <block type="defect"></block>
        <block type="cooperate"></block>
        <block type="if_last_move_cooperate"></block>
        <block type="if_last_move_defect"></block>
    </xml>
=======
    <xml id="toolbox" style="display: none">
        <block type="defect"></block>
        <block type="cooperate"></block>
        <block type="if_last_move_cooperate"></block>
        <block type="if_last_move_defect"></block>
        <block type="first_move"></block>
    </xml>
>>>>>>> REPLACE
```

This will add the 'first_move' block to the toolbox of the Blockly interface.