aider: Added a loop block to the custom blocks and added it to the toolbox in the index.html file.
# Aider chat conversation:

USER: make a loop block in which we can plugin the rules for subsequent mvoes in the game

ASSISTANT: To create a loop block, we need to modify the `client/custom_blocks.js` file to define the new block. The loop block will have an input statement where we can plug in the rules for subsequent moves in the game.

Here is the change we need to make:

```javascript
client/custom_blocks.js
<<<<<<< SEARCH
  {
    "type": "first_move",
    "message0": "First move %1",
    "args0": [
      {
        "type": "input_value",
        "name": "MOVE",
        "check": ["defect", "cooperate"]
      }
    ],
    "previousStatement": null,
    "nextStatement": null,
    "colour": 90,
  }
]);
=======
  {
    "type": "first_move",
    "message0": "First move %1",
    "args0": [
      {
        "type": "input_value",
        "name": "MOVE",
        "check": ["defect", "cooperate"]
      }
    ],
    "previousStatement": null,
    "nextStatement": null,
    "colour": 90,
  },
  {
    "type": "loop",
    "message0": "Loop %1",
    "args0": [
      {
        "type": "input_statement",
        "name": "STATEMENTS"
      }
    ],
    "previousStatement": null,
    "nextStatement": null,
    "colour": 20,
  }
]);
>>>>>>> REPLACE
```

After defining the new block, we need to add it to the toolbox in the `client/index.html` file:

```html
client/index.html
<<<<<<< SEARCH
    <xml id="toolbox" style="display: none">
        <block type="defect"></block>
        <block type="cooperate"></block>
        <block type="if_last_move_cooperate"></block>
        <block type="if_last_move_defect"></block>
        <block type="first_move"></block>
    </xml>
=======
    <xml id="toolbox" style="display: none">
        <block type="defect"></block>
        <block type="cooperate"></block>
        <block type="if_last_move_cooperate"></block>
        <block type="if_last_move_defect"></block>
        <block type="first_move"></block>
        <block type="loop"></block>
    </xml>
>>>>>>> REPLACE
```

These changes will create a new loop block and add it to the toolbox. You can then use this block to define the rules for subsequent moves in the game.