aider: Add a `payoff` variable to the `env` object in `game_logic.js`.
# Aider chat conversation:

USER: we also need a payoff variable let us call it payoff

ASSISTANT: Sure, we can add a `payoff` variable to the `env` object in the `game_logic.js` file. This variable can be used to store the payoff for each player after each round of the game. Here's how we can do it:

```javascript
server/game_logic.js
<<<<<<< SEARCH
    env: {
        // The global state of the game
        players: [], // An array to store the players
        moves: [], // An array to store the moves of the players
        gameMatrix: { // The game matrix
            'cooperate': {
                'cooperate': [2, 2],
                'defect': [0, 3]
            },
            'defect': {
                'cooperate': [3, 0],
                'defect': [1, 1]
            }
        },
        numberOfRounds: 0, // The number of rounds in the game
    },
=======
    env: {
        // The global state of the game
        players: [], // An array to store the players
        moves: [], // An array to store the moves of the players
        gameMatrix: { // The game matrix
            'cooperate': {
                'cooperate': [2, 2],
                'defect': [0, 3]
            },
            'defect': {
                'cooperate': [3, 0],
                'defect': [1, 1]
            }
        },
        numberOfRounds: 0, // The number of rounds in the game
        payoff: [0, 0], // The payoff for each player
    },
>>>>>>> REPLACE
```

This change adds a `payoff` property to the `env` object, which is an array of two elements. The first element represents the payoff for player 1, and the second element represents the payoff for player 2. Initially, both payoffs are set to 0.