aider: Added a `runGame` function that resets the game, executes the code of the two players, records their moves, and updates the payoff according to the game matrix.
# Aider chat conversation:

USER: let us write the runGame function; it needs to have arguments code1 and code2 which store the code coming from the blocks of the two clients and it needs to modify the global game state, in particular the payoff

ASSISTANT: Sure, we can add a `runGame` function to the `game_logic.js` file. This function will take two arguments, `code1` and `code2`, which represent the code coming from the blocks of the two clients. The function will modify the global game state, in particular the `payoff` property of the `env` object.

Here's how we can do it:

```javascript
server/game_logic.js
<<<<<<< SEARCH
    // A function to reset the game
    reset: function() {
        this.env.players = [];
        this.env.moves = [];
    }
=======
    // A function to reset the game
    reset: function() {
        this.env.players = [];
        this.env.moves = [];
        this.env.payoff = [0, 0];
    },

    // A function to run the game
    runGame: function(code1, code2) {
        // Reset the game
        this.reset();

        // Execute the code of the two players
        let move1 = eval(code1);
        let move2 = eval(code2);

        // Record the moves
        this.recordMove(1, move1);
        this.recordMove(2, move2);

        // Update the payoff
        let payoff1 = this.env.gameMatrix[move1][move2][0];
        let payoff2 = this.env.gameMatrix[move1][move2][1];
        this.env.payoff[0] += payoff1;
        this.env.payoff[1] += payoff2;
    }
>>>>>>> REPLACE
```

This change adds a `runGame` function that resets the game, executes the code of the two players, records their moves, and updates the payoff according to the game matrix. The `reset` function is also modified to reset the payoff to 0 for both players at the start of the game.

Please note that the `eval` function is used to execute the code of the players. This function can execute arbitrary JavaScript code, which can be a security risk if the code comes from an untrusted source. In a real-world application, you should use a safer method to execute the code, such as running it in a sandbox or using a safe JavaScript interpreter.