aider: Add generator function for "first move" block to check if the current round is the first round and return the move specified in the block.
# Aider chat conversation:

USER: next we generate code for "first move" it should be sth like "if round=0 then <whatever the argument is>"

ASSISTANT: To generate the code for the "first move" block, we need to modify the `client/custom_blocks.js` file. We will add a new generator function for the "first move" block. This function will generate JavaScript code that checks if the current round is the first round (round 0), and if so, it will return the move specified in the block.

Here is the *SEARCH/REPLACE block* for this change:

```javascript
client/custom_blocks.js
<<<<<<< SEARCH
Blockly.JavaScript['cooperate'] = function(block) {
  return ['"cooperate"', Blockly.JavaScript.ORDER_ATOMIC];
};
=======
Blockly.JavaScript['cooperate'] = function(block) {
  return ['"cooperate"', Blockly.JavaScript.ORDER_ATOMIC];
};

Blockly.JavaScript['first_move'] = function(block) {
  var move = Blockly.JavaScript.valueToCode(block, 'MOVE', Blockly.JavaScript.ORDER_ATOMIC);
  var code = 'if (round === 0) { return ' + move + '; }';
  return code;
};
>>>>>>> REPLACE
```

This change adds a new generator function for the "first move" block. This function generates JavaScript code that checks if the current round is the first round (round 0), and if so, it returns the move specified in the block. The move is obtained from the block's "MOVE" input value.