aider: Added generator function for the "loop" block to check if the current round is greater than 0 and execute the statements.
# Aider chat conversation:

USER: looks good, now let us do the loop it is sth if "if round>0 ..."

ASSISTANT: To generate the code for the "loop" block, we need to modify the `client/custom_blocks.js` file. We will add a new generator function for the "loop" block. This function will generate JavaScript code that checks if the current round is greater than 0, and if so, it will execute the statements specified in the block.

Here is the *SEARCH/REPLACE block* for this change:

```javascript
client/custom_blocks.js
<<<<<<< SEARCH
Blockly.JavaScript['first_move'] = function(block) {
  var move = Blockly.JavaScript.valueToCode(block, 'MOVE', Blockly.JavaScript.ORDER_ATOMIC);
  var code = 'if (round === 0) { return ' + move + '; }';
  return code;
};
=======
Blockly.JavaScript['first_move'] = function(block) {
  var move = Blockly.JavaScript.valueToCode(block, 'MOVE', Blockly.JavaScript.ORDER_ATOMIC);
  var code = 'if (round === 0) { return ' + move + '; }';
  return code;
};

Blockly.JavaScript['loop'] = function(block) {
  var statements = Blockly.JavaScript.statementToCode(block, 'STATEMENTS');
  var code = 'if (round > 0) { ' + statements + ' }';
  return code;
};
>>>>>>> REPLACE
```

This change adds a new generator function for the "loop" block. This function generates JavaScript code that checks if the current round is greater than 0, and if so, it executes the statements specified in the block. The statements are obtained from the block's "STATEMENTS" input value.