aider: Added generator function for "if_last_move_cooperate" block.
# Aider chat conversation:

USER: "if last move was cooperate" ...

ASSISTANT: To generate the code for the "if_last_move_cooperate" block, we need to modify the `client/custom_blocks.js` file. We will add a new generator function for the "if_last_move_cooperate" block. This function will generate JavaScript code that checks if the last move of the opponent was to cooperate, and if so, it will return the move specified in the block.

Here is the *SEARCH/REPLACE block* for this change:

```javascript
client/custom_blocks.js
<<<<<<< SEARCH
Blockly.JavaScript['loop'] = function(block) {
  var statements = Blockly.JavaScript.statementToCode(block, 'STATEMENTS');
  var code = 'if (round > 0) { ' + statements + ' }';
  return code;
};
=======
Blockly.JavaScript['loop'] = function(block) {
  var statements = Blockly.JavaScript.statementToCode(block, 'STATEMENTS');
  var code = 'if (round > 0) { ' + statements + ' }';
  return code;
};

Blockly.JavaScript['if_last_move_cooperate'] = function(block) {
  var move = Blockly.JavaScript.valueToCode(block, 'MOVE', Blockly.JavaScript.ORDER_ATOMIC);
  var code = 'if (getLastMove(opponent) === "cooperate") { return ' + move + '; }';
  return code;
};
>>>>>>> REPLACE
```

This change adds a new generator function for the "if_last_move_cooperate" block. This function generates JavaScript code that checks if the last move of the opponent was to cooperate, and if so, it returns the move specified in the block. The move is obtained from the block's "MOVE" input value.